Fix warnings.
In #785 the type of YR_MEMORY_BLOCK's "size" and "base" fields was changed from size_t to uint64_t. This changed introduced a number of warnings due to loss of precision when forcing the uint64_t into a size_t.

+++ libyara/compiler.c
@@ -110,7 +110,7 @@ const char* _yr_compiler_default_include_callback(
   file_size = stbuf.st_size;
   #endif
 
-  file_buffer = (char*) yr_malloc(file_size + 1);
+  file_buffer = (char*) yr_malloc((size_t) file_size + 1);
 
   if (file_buffer == NULL)
   {
@@ -123,7 +123,7 @@ const char* _yr_compiler_default_include_callback(
     return NULL;
   }
 
-  if (file_size != read(fd, file_buffer, file_size))
+  if (file_size != read(fd, file_buffer, (size_t) file_size))
   {
     yr_free(file_buffer);
