+++ libyara/include/yara/exefiles.h
@@ -38,6 +38,6 @@ uint64_t yr_get_entry_point_offset(
 uint64_t yr_get_entry_point_address(
     const uint8_t* buffer,
     size_t buffer_length,
-    size_t base_address);
+    uint64_t base_address);
 
 #endif
